move common code to a concern
